+++ novawallet/Modules/Swaps/Setup/SwapSetupProtocols.swift
@@ -51,7 +51,7 @@ protocol SwapSetupInteractorOutputProtocol: AnyObject {
     func didReceive(balance: AssetBalance?, for chainAsset: ChainAssetId, accountId: AccountId)
 }
 
-protocol SwapSetupWireframeProtocol: AnyObject, AlertPresentable, CommonRetryable, ErrorPresentable {
+protocol SwapSetupWireframeProtocol: AnyObject, AlertPresentable, CommonRetryable, ErrorPresentable, ShortTextInfoPresentable {
     func showPayTokenSelection(
         from view: ControllerBackedProtocol?,
         chainAsset: ChainAsset?,
