+++ novawallet/en.lproj/Localizable.strings
@@ -1394,6 +1394,7 @@
 "swaps.setup.slippage.error.amount.bounds" = "Enter a value between %@ and %@";
 "swaps.setup.slippage.warning.low.amount" = "Transaction might be reverted because of low slippage tolerance.";
 "swaps.setup.slippage.warning.high.amount" = "Transaction might be frontrun because of high slippage.";
+"swaps.setup.slippage.description" = "Swap slippage is a common occurrence in decentralized trading where the final price of a swap transaction might slightly differ from the expected price, due to changing market conditions.";
 "swaps.rate.description" = "Exchange rate between two different cryptocurrencies. It represents how much of one cryptocurrency you can get in exchange for a certain amount of another cryptocurrency.";
 "swaps.network.fee.description" = "A network fees charged by the blockchain to process and validate any transactions. May vary depending on network conditions or transaction speed.";
 "common.alert.external.link.disclaimer.title" = "Continue in browser?";
