+++ novawallet/ru.lproj/Localizable.strings
@@ -1394,6 +1394,7 @@
 "swaps.setup.slippage.error.amount.bounds" = "Введите значение между %@ и %@";
 "swaps.setup.slippage.warning.low.amount" = "Транзакция может быть отменена из-за низкого значения проскальзывания.";
 "swaps.setup.slippage.warning.high.amount" = "Транзакция может быть приостановлена из-за высокого значения проскальзывания.";
+"swaps.setup.slippage.description" = "Обменное проскальзывание - обычное явление в децентрализованной торговле, где окончательная цена транзакции обмена может незначительно отличаться от ожидаемой цены из-за изменения рыночных условий.";
 "swaps.rate.description" = "Обменный курс между двумя различными криптовалютами. Он представляет, сколько одной криптовалюты вы можете получить в обмен на определенное количество другой криптовалюты.";
 "swaps.network.fee.description" = "Это комиссия сети, взимаемая блокчейном за обработку и подтверждение любых транзакций. Она может изменяться в зависимости от условий в сети или скорости выполнения транзакции.";
 "common.alert.external.link.disclaimer.title" = "Продолжить в браузере?";
